fix some mistakes finish Traversal file and dirs func Signed-off-by: 921 Huang <yenan.huang@kcl.ac.uk>
