+++ desktop/Meerkats_Win/Class/func.cs
@@ -1,0 +1,63 @@
+﻿using System;
+using System.Collections.Generic;
+using System.IO;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Meerkats_Win.Class
+{
+    class func
+    {
+        // stringhex(md5) => byte[]
+        public byte[] HexStrTobyte(string hexString)
+        {
+            try
+            {
+                hexString = hexString.Replace(" ", "");
+                if ((hexString.Length % 2) != 0)
+                    hexString += " ";
+                byte[] returnBytes = new byte[hexString.Length / 2];
+                for (int i = 0; i < returnBytes.Length; i++)
+                    returnBytes[i] = Convert.ToByte(hexString.Substring(i * 2, 2).Trim(), 16);
+                return returnBytes;
+
+                ////byte[] => stringhex(md5)
+
+                //byte[] buffer = {};
+
+                //StringBuilder strBuider = new StringBuilder();
+                //for (int index = 0; index < count; index++)
+                //{
+                //    strBuider.Append(((int)buffer[index]).ToString("X2"));
+                //}
+            }
+            catch (Exception ex)
+            {
+                throw new Exception(ex.Message);
+            }
+        }
+
+        public string GetMD5HashFromFile(string fileName)
+        {
+            try
+            {
+                FileStream file = new FileStream(fileName, FileMode.Open);
+                System.Security.Cryptography.MD5 md5 = new System.Security.Cryptography.MD5CryptoServiceProvider();
+                byte[] retVal = md5.ComputeHash(file);
+                file.Close();
+
+                StringBuilder sb = new StringBuilder();
+                for (int i = 0; i < retVal.Length; i++)
+                {
+                    sb.Append(retVal[i].ToString("x2"));
+                }
+                return sb.ToString();
+            }
+            catch (Exception ex)
+            {
+                throw new Exception("GetMD5HashFromFile() fail,error:" + ex.Message);
+            }
+        }
+    }
+}
